requirements.txt
anyio==3.4.0
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
fastapi==0.70.1
h11==0.12.0
html5lib==1.1
idna==3.3
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
webencodings==0.5.1
